change dateTimeOfActivity to double (GQL Float is Java double)

+++ backend/src/main/java/com/team3/backend/datafetchers/log/LogDataFetcher.java
@@ -31,7 +31,7 @@ public class LogDataFetcher {
 
     public DataFetcher createLog() {
         return dataFetchingEnvironment -> {
-            int dateTimeOfActivity = dataFetchingEnvironment.getArgument("dateTimeOfActivity");
+            double dateTimeOfActivity = dataFetchingEnvironment.getArgument("dateTimeOfActivity");
             String notes = dataFetchingEnvironment.getArgument("notes");
             List<String> categories = dataFetchingEnvironment.getArgument("categories");
             List<String> mood = dataFetchingEnvironment.getArgument("mood");
